Create Dockerfile
The self hosting guide linked below directs users to create this file, but lines 12 and 13 on the guide lack the "ENV " which results in error code 12. This docker file would address that and mitigate the need to create the file manually to begin with.

https://help.ente.io/self-hosting/guides/web-app

I'm fairly new at contributing. If there's a better way to bring awareness to this issue, I would be happy to go about doing it that way.